#1905 - Refactor Render class: convert raphaelRender.js to TypeScript and add necessary types * Refactor raphaelRender.js, draw.js to TypeScript and provide correct types * Fix TS errors after the refactoring
