+++ packages/ketcher-core/src/application/editor/shared/constants.ts
