+++ packages/ketcher-core/src/application/editor/shared/utils.ts
@@ -14,34 +14,35 @@
  * limitations under the License.
  ***************************************************************************/
 
-import { Vec2 } from 'domain/entities'
+import { Struct, Vec2 } from 'domain/entities'
 import { inRange } from 'lodash'
+import { BondAtoms } from 'application/editor/shared/utils.types'
 
 let FRAC = Math.PI / 12 // '15º'
 
-function setFracAngle(angle) {
+function setFracAngle(angle: number): void {
   FRAC = (Math.PI / 180) * angle
 }
 
-function calcAngle(pos0, pos1) {
+function calcAngle(pos0: Vec2, pos1: Vec2): number {
   const v = Vec2.diff(pos1, pos0)
   return Math.atan2(v.y, v.x)
 }
 
-function fracAngle(angle, angle2) {
+function fracAngle(angle, angle2): number {
   if (angle2) angle = calcAngle(angle, angle2)
   return Math.round(angle / FRAC) * FRAC
 }
 
-function calcNewAtomPos(pos0, pos1, ctrlKey) {
-  const v = new Vec2(1, 0).rotate(
+function calcNewAtomPos(pos0: Vec2, pos1: Vec2, ctrlKey: boolean): Vec2 {
+  const vector = new Vec2(1, 0).rotate(
     ctrlKey ? calcAngle(pos0, pos1) : fracAngle(pos0, pos1)
   )
-  v.add_(pos0) // eslint-disable-line no-underscore-dangle
-  return v
+  vector.add_(pos0) // eslint-disable-line no-underscore-dangle
+  return vector
 }
 
-function degrees(angle) {
+function degrees(angle: number): number {
   let degree = Math.round((angle / Math.PI) * 180)
   if (degree > 180) degree -= 360
   else if (degree <= -180) degree += 360
@@ -51,11 +52,17 @@ function degrees(angle) {
 const BONDS_MERGE_ANGLE = 10 // 'º'
 const BONDS_MERGE_SCALE = 0.2
 
-function mergeBondsParams(struct1, bond1, struct2, bond2) {
-  const begin1 = struct1.atoms.get(bond1.begin)
-  const begin2 = struct2.atoms.get(bond2.begin)
-  const end1 = struct1.atoms.get(bond1.end)
-  const end2 = struct2.atoms.get(bond2.end)
+function mergeBondsParams(
+  struct1: Struct,
+  bond1: BondAtoms,
+  struct2: Struct,
+  bond2: BondAtoms
+) {
+  // TODO find a solution to remove not-null assertion from result atoms below
+  const begin1 = struct1.atoms.get(bond1.begin)!
+  const begin2 = struct2.atoms.get(bond2.begin)!
+  const end1 = struct1.atoms.get(bond1.end)!
+  const end2 = struct2.atoms.get(bond2.end)!
 
   const angle = calcAngle(begin1.pp, end1.pp) - calcAngle(begin2.pp, end2.pp)
   const mergeAngle = Math.abs(degrees(angle) % 180)
