+++ packages/ketcher-core/src/application/editor/shared/utils.types.ts
@@ -1,0 +1,3 @@
+import { Bond } from 'domain/entities'
+
+export type BondAtoms = Pick<Bond, 'begin' | 'end'>
