+++ packages/ketcher-core/src/application/render/raphael-ext.ts
@@ -20,14 +20,15 @@ import Raphael from 'raphael'
 import { Vec2 } from 'domain/entities'
 
 // TODO: refactor ugly prototype extensions to plain old functions
-Raphael.el.translateAbs = function (x, y) {
+Raphael.el.translateAbs = function (x: number, y: number): void {
   this.delta = this.delta || new Vec2()
+  // TODO check that only numbers might be passed to this function
   this.delta.x += x - 0
   this.delta.y += y - 0
   this.transform('t' + this.delta.x.toString() + ',' + this.delta.y.toString())
 }
 
-Raphael.st.translateAbs = function (x, y) {
+Raphael.st.translateAbs = function (x: number, y: number): void {
   this.forEach((el) => {
     el.translateAbs(x, y)
   })
