+++ packages/ketcher-core/src/application/render/restruct/reloop.js
@@ -19,9 +19,7 @@ import { Bond, Vec2 } from 'domain/entities'
 import { LayerMap } from './generalEnumTypes'
 import ReObject from './reobject'
 import { Scale } from 'domain/helpers'
-import util from '../util'
-
-const tfx = util.tfx
+import { tfx } from 'utilities'
 
 class ReLoop extends ReObject {
   constructor(loop) {
