+++ packages/ketcher-core/src/application/render/restruct/restruct.ts
@@ -308,7 +308,7 @@ class ReStruct {
     })
   }
 
-  getVBoxObj(selection?): Box2Abs | null {
+  getVBoxObj(selection?): Box2Abs {
     selection = selection || {}
 
     if (isSelectionEmpty(selection)) {
