+++ packages/ketcher-core/src/utilities/tfx.ts
@@ -1,3 +1,4 @@
+// TODO add tests and rename to toFixed
 export function tfx<TValue extends number | string>(value: TValue): string {
   let parsedValue: number
   if (typeof value === 'number') {
