+++ packages/ketcher-react/src/script/editor/Editor.ts
@@ -503,11 +503,22 @@ class Editor implements KetcherEditor {
   findItem(event: any, maps: any, skip: any = null) {
     const pos = new Vec2(this.render.page2obj(event))
 
-    return closest.item(this.render.ctab, pos, maps, skip, this.render.options)
+    return closest.item(
+      this.render.ctab,
+      pos,
+      maps,
+      skip,
+      this.render.options.scale
+    )
   }
 
   findMerge(srcItems: any, maps: any) {
-    return closest.merge(this.render.ctab, srcItems, maps, this.render.options)
+    return closest.merge(
+      this.render.ctab,
+      srcItems,
+      maps,
+      this.render.options.scale
+    )
   }
 
   explicitSelected() {
