+++ packages/ketcher-react/src/script/editor/tool/template.ts
@@ -153,7 +153,7 @@ class TemplateTool {
       })
   }
 
-  async mousedown(event) {
+  async mousedown(event: MouseEvent) {
     this.event = event
 
     if (this.functionalGroups.size) {
@@ -187,7 +187,7 @@ class TemplateTool {
     this.editor.hover(null)
 
     this.dragCtx = {
-      xy0: this.editor.render.page2obj(this.event),
+      xy0: this.editor.render.page2obj(event),
       item: this.editor.findItem(this.event, this.findItems)
     }
